chore: 리뷰 반영하여 수정
