+++ src/pages/NaverPage/components/index.ts
@@ -1,2 +1,5 @@
 export { default as HideButton } from './HideButton'
 export { default as RestAreaInfo } from './RestAreaInfo'
+export { default as InputSubmit } from './InputSubmit'
+export { default as Logos } from './Logos'
+export { default as PathInfo } from './PathInfo'
